server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic-settings==2.6.0
psycopg[binary]==3.2.3
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
httpx==0.27.2
